aider: Added a new function `highlightTags` in DataTable.js to wrap tags in a span with a specific class and modified the rendering of the notes to use this function, also added CSS rules for the `.tag` class in DataTable.css.
# Aider chat conversation:

USER: I did `git reset --hard HEAD~1` to discard the last edits.

ASSISTANT: Alright. If you need any further assistance or want to make more changes, feel free to ask.

USER: In the Note column, all words starting with "#" should have a nice pastel rounded background, those are called tags. Make a separate function in DataTabel.js to highlight tags. Remember that tag is a word, and it can be followed by a coma, dot, etc. and also it can be the last word in the description. Make sure the function returns proper values.

ASSISTANT: To highlight the tags in the Note column, we need to create a new function in the `DataTable.js` file that wraps the tags (words starting with '#') in a span with a specific class. We also need to modify the way the notes are rendered to use this new function.

Let's start by creating the new function:

```javascript
www/app/src/DataTable.js
<<<<<<< SEARCH

=======
function highlightTags(note) {
    return note.split(' ').map((word, index) =>
        word.startsWith('#') ?
            <span key={index} className="tag">{word}</span> :
            word
    ).join(' ');
}
>>>>>>> REPLACE
```

Then, let's modify the rendering of the notes to use this new function:

```javascript
www/app/src/DataTable.js
<<<<<<< SEARCH
<td>{transaction.note}</td>
=======
<td>{highlightTags(transaction.note)}</td>
>>>>>>> REPLACE
```

Finally, let's add the CSS rules for the `.tag` class in the `DataTable.css` file:

```css
www/app/src/DataTable.css
<<<<<<< SEARCH

=======
.tag {
    background-color: #f2f2f2; /* pastel color */
    border-radius: 5px; /* rounded corners */
    padding: 2px 5px; /* some padding */
}
>>>>>>> REPLACE
```

In the `REPLACE` section, I've added a new `.tag` class with a pastel background color, rounded corners, and some padding. This will style the tags in the Note column.