aider: Refactored the `highlightTags` function to use a regular expression to match tags instead of splitting the string.
# Aider chat conversation:

USER: This function does not take into consideration, that tag might be the last word in the string (when split(' ') will not work. also tag can be followed by any puctuation. use a regex to find tags and not a split (' ')

ASSISTANT: You're right, using `split(' ')` won't handle cases where the tag is the last word in the string or when it's followed by punctuation. We can use a regular expression to match the tags instead. Here's how we can modify the `highlightTags` function:

```javascript
www/app/src/DataTable.js
<<<<<<< SEARCH
function highlightTags(note) {
    return note.split(' ').map((word, index) =>
        word.startsWith('#') ?
            <span key={index} className="tag">{word}</span> :
            word
    ).join(' ');
}
=======
function highlightTags(note) {
    return note.replace(/(#\w+)/g, '<span class="tag">$1</span>');
}
>>>>>>> REPLACE
```

In the `REPLACE` section, I've replaced the `split(' ')` method with a regular expression that matches any word starting with '#'. The `replace` method then wraps these matches in a span with the `tag` class. This will handle tags that are the last word in the string or followed by punctuation.